+++ aioaerospike/client.py
@@ -3,7 +3,7 @@ from functools import wraps
 from typing import Any, Dict, List, Optional
 
 from .protocol.general import AerospikeHeader, AerospikeMessage
-from .protocol.message import delete_key, get_key, put_key
+from .protocol.message import delete_key, get_key, key_exists, put_key
 
 
 class AerospikeClientNotConnected(Exception):
@@ -100,3 +100,18 @@ class AerospikeClient:
             raise Exception(
                 f"Unexpected result code {response.message.result_code}"
             )
+
+    @require_connection
+    async def key_exists(self, namespace: str, set_name: str, key: str) -> bool:
+        message = key_exists(namespace, set_name, key)
+        data = AerospikeMessage(message).pack()
+        self._writer.write(data)
+        await self._writer.drain()
+        response = await self._get_response()
+        if response.message.result_code == 2:
+            return False
+        elif response.message.result_code != 0:
+            raise Exception(
+                f"Unexpected result code {response.message.result_code}"
+            )
+        return True