+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.4 (XXXX-XX-XX)
 - Added delete key method
+- Added key_exists method
 
 ## 0.1.3 (2019-12-07) (Unreleased)
 - Changed all enums to uppercase